[REVERT ME] rril: add an oem hook command for AT+NEER (AT&T support)
BZ: 92122

To support AT&T network, it's necessary for the RRIL to provide a way
to have the extended error of the last registration (AT+NEER).

This patch updates the code to retrieve the error code in the AT+NEER
response (returned by #<code>) and returns it via an OEM hook.

Change-Id: I5f681f74b6ddb0dd32ed47315d186a560648daec
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>