[REVERT ME] rril: add an oem hook command for AT+NEER (AT&T support)
BZ: 92122

To support AT&T network, it's necessary for the RRIL to provide a way
to have the extended error of the last registration (AT+NEER).

This patch updates the code to retrieve the error code in the AT+NEER
response (returned by #<code>) and returns it via an OEM hook.

Change-Id: I5f681f74b6ddb0dd32ed47315d186a560648daec
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/99477
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Morel, YannickX <yannickx.morel@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>