rril: update channel assignement for call operations
BZ: 96001

To create a new call with an active call ongoing, the Android/FW
send the REQUEST_SWITCH_WAITING_OR_HOLDING_AND_ACTIVATE request
first but it doesn't wait the completion of this one before to
sent the DIAL request. If those operations are did on different
channel at RRIL level, we can have some error at modem level.

To avoid modem issue, the call operations must be done on the
same channel in the goal to serialize them. This patch updates
the channel assignement to have this.

Change-Id: Icdcc488411236a3b1d72d700df0134c136fcf0dd
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/99682
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>