rril: Fix rild socket closed due to modem events
BZ: 98152

In flight mode, if there are no clients for modem,
then modem is switched off. If the Wifi state is
toggled in modem off in flight mode, modem is
powered on/off respectively. This results in rild
socket disconnect error.

Following changes are done to address this issue:
- Check the modem status before opening each port.
- Clear the command queue and free the command's
context data if the modem is not up.
- Remove the CloseChannelPorts call done on
MODEM_DOWN as this has been already taken care
in the respective modem events
- Incase of modem recovery request, set the
initialization as unsuccessful before closing the
channel ports.

Change-Id: Ic9a1194b17bb9e4072ddb4b16e41200ca31b9369
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/100777
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Amegah, AndreX <andrex.amegah@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>