rril: Fix first DTMF request ignored on some cases
BZ: 96043

When the user accepts the waiting call, current
call is put on hold and the waiting call is accepted.
When the user accepts the waiting call, DTMF state
is set to E_DTMF_STATE_FLUSH to flush all the pending
DTMF requests. This is done even when there is no
pending DTMF requests. So, when user sends DTMF
requests for the second call, DTMF state is still
in E_DTMF_STATE_FLUSH resulting in first DTMF request
getting ignored.

Following changes are done to address this issue
- Call List is cached in the rapid ril
- Current call id is maintained for each DTMF requests
- DTMF request is allowed based on current dtmf and call
state.
- Remove E_DTMF_STATE_FLUSH as this has been replaced
by DtmfAllowed flag which is maintained for each call.

Change-Id: I01c2ba3c278f347d50b04c60198116b5af2e56f5
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/101226
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>