[PORT FROM R42S]rril: Fix issues on AT+CCHO, AT+CCHC and AT+CGLA commands
BZ: 101367

AT+CGLA command: the RRIL should add a \r character at the end
of the command. Not doing this leads to a transmit timeout.

AT+CCHC command: the function which builds this command should return
the value RRIL_RESULT_OK at the end. Not doing this prevent the
command from being effectively sent.

AT+CCHO command: the function in charge of parsing the answer to
this command should be able to skip the +CCHO string if present
in the answer.

Change-Id: I6672e715d985a3f9661cfa6daf74192351f9b847
Orig-Change-Id: I86c49691e6c29740a6023c92b739aabe3fc182fe
Signed-off-by: Sebastien Griffoul <sebastien.griffoul@intel.com>
Reviewed-on: http://android.intel.com:8080/102917
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Niciarz, Sebastian <sebastian.niciarz@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>