rril: Handle platform shutdown cleanly
BZ: 102072

Upon platform shutdown, rapid ril requests mmgr for
force shutdown and completes the request immediately.
Upon MODEM_DOWN event from MMGR, rapid ril sets the
radio state to RADIO_OFF and notifies telephony
framework of the radio state change. Eventhough,
this is correct but it is not clean which is resulting
in abnormal shutdown.

Following changes are done to address the issue:

- Send AT+CHLD=8;+CGATT=0 on platform shutdown request.
Upon response of this command, close channel ports and
request MMGR for electrical power down of modem.
Upon MODEM_DOWN event, notify of radio state change and
also complete platform
shutdown request.
- Remove CloseChannelPorts done on reception of modem
cold reset/warm reset/shutdown etc as this is causing
rild socket getting closed. As the channel response
thread is blocked forever, closing of the ports in
an another thread is resulting in rild socket getting
closed. Note: Incase of platform shutdown, channel
ports are closed before requesting mmgr for shutdown.
- Introduce radio off reason, radio power constants
- Start the rapid ril with modem state set to
MODEM_STATE_UNKNOWN.
- Rename ModemPowerOnEvent to RadioPoweredOnEvent
- Handling of RIL_REQUEST_RADIO_POWER based on
modem state is moved to CoreRadioPower function.
- Modem specific classes will provide the commands
for RADIO_POWER request based on modem state and
requested radio state.
- Ignore modem recovery request during platform
shutdown.
- If platform shutdown request fails, then set
the radio state to off and comple the pending
RADIO_POWER request.
- New function SetRadioStateAndNotify is added
to CRadioState class. This function needs to be
used if the radio state needs to be set and also
to be notified to framework.

Change-Id: I43cedd36309cb1af53a9730176d5eef0d1aaedf4
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/103810
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>