rril: Dont destroy system manager on modem events
BZ: 103189

Upon modem events like out of service, platform
reboot, platform shutdown system manager is deleted.
This results in CTE object also deleted. So, when
there is a ril request, rapid ril calls GetTE
which calls exit as there is no terminal instance.

With this patch, system manager will not be
deleted on modem events. Also, request to power
on/off the radio will be rejected if the modem
state is out of service or platform reboot.

Change-Id: If5d45b4e308e8fb9712273655c26fb5f0529b31d
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/104521
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Berat, FredericX <fredericx.berat@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>