[PORT FROM R42S] rril: stop channel threads before closing ports
BZ: 103970

Channel threads(command and response thread) needs
to be stopped before closing the channel ports.
This is done by Signal the cancel event and also
by calling ThreadManager's Stop function. Response
thread is not stopped as it is polling for events
in the port descriptor.

With this patch, a pipe is created. Response thread
will poll for events on the port and also pipe file
descriptor. If the response thread needs to be exited,
then data will be written to the write pipe descriptor.
Upon events on the pipe descriptor, response thread
will stop polling for data and exits.

Note: As the response thread is stopped properly with
this patch, CloseChannelPorts done on MODEM_DOWN, SHUTDOWN,
WARM RESET, COLD RESET events is restored.

Change-Id: I1da47ed8c3c832c9122f1a0d1f884bad2bfa8879
Orig-Change-Id: I22bcc854741f3d98c1b1f083831ed2e08c6b2d4d
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/105987
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>