[PORT FROM R42S] rril: handle EAGAIN and EINTR in TTY opening code
BZ: 105638

[WORKAROUND] TTY opening code does not handle EAGAIN and EINTR error
codes meaning that back-to-back close / open may fail due to the time
it takes for the close to actually happen on the MUX control channel
(as close is done on a non-blocking file descriptor, the code may try
re-opening a TTY while the close is still in progress).

Set as WORKAROUND while investigating a cleaner solution (like
putting the file descriptor as BLOCKING before the close call to
ensure proper fd closing).

Change-Id: I99e1b48700a1cfac2fad13c1215e01f9d850bf3a
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Reviewed-on: http://android.intel.com:8080/107655
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>