rril: manage registration reject cause
BZ: 62327

Registration reject cause is not notified
to framework.

If the registration is rejected, then rapid ril
needs to send the reject cause as part of the
response to RIL_REQUEST_VOICE_REGISTRATION_STATE
and L_REQUEST_DATA_REGISTRATION_STATE.E.
Currently, this is not done in rapid ril.

As per 3GPP 27.007 Release 11, if CREG command
is activated with <n>=3, then registration
reject cause is sent if available.

Following changes are done to fix the issue
- Active registration status reporting with
<n> value based on modem type.
- If the registration status is denied but
if the reject cause is not available, then
clear the cache flag to allow the registration
status request to be queried from the modem
directly.
- If parsing of LAC, CID fails then log the
error but continue parsing the response.
- Fixes the compilation warning due to CTE member
variable m_uiAct not initialized in correct order.

Change-Id: I08b24d32dfdb9115f0bb8e8ab7f0611c88e803e2
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/107310
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>