rril: Fix issue in notifying radio off state
BZ: 110148

When RADIO_POWER off is received with no reason and
in modem up state, request is handled successfully
but the radio state change notification is not sent
to framework. Due to this, device status in not
registered forever.

When RADIO_POWER off is received during MODEM_UP
but with no radio off reason, set and notify the
radio state change to framework.

Change-Id: Ia698d93b2a2beeef5b9ef9e4eb0a795345d028f8
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/109733
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>