rril: Update sim app state as unknown in radio off
BZ: 107119

Upon airplane mode activation request, rapid ril
will release the mmgr resource. If there is no
more modem clients, mmgr will power off the modem.
If the airplane mode is activated when the sim
pin dialog is shown, then the SIM pin dialog
is not dismissed on radio powered down. So, even
if user enters the correct sim pin, SIM pin dialog
won't be dismissed as the request can't be sent
to modem in powered off state.

Complete GET_SIM_STATUS with sim state as present
and app state set as unknown. This is needed to
dismiss the sim pin dialog when the sim is not ready.

Change-Id: I3543bde0ce3e5259dd733676ac39a0188b810400
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/110418
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Butscher, MorganeX <morganex.butscher@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>