rril: Remove modem recovery done on MODEM_UP event
BZ: 113030

If the E_MMGR_EVENT_MODEM_UP event is received
and if the previous state is MODEM_STATE_UNKNOWN,
modem recovery is done which is not needed.

As the closing of channels is now done properly and
also opening of ports is delayed until the close port
is completed, no need to do the modem recovery.

Change-Id: I632fb4f07856f1a2c60f249751e2d45125099e06
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/111434
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>