rril: fix issue in call drop reporting for mobile release case
BZ: 113690

When a call drop is due to a mobile release, the call drop
reporting under crashtool should set "MOBILE RELEASE" in the
data 0 field.

Due to an issue in the code, the data 0 is an empty string with
the current code. This patch fixes the issue in the RRIL code
to have the correct reporting.

Change-Id: Id8c90b11b786191adb573bc2f09df970caaa5cf3
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/111722
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>