rril: Fix issue in parsing CGCONTRDP
BZ: 114099

In case of dual stack capability, parsing
of CGCONTRDP is not done correctly.

With this patch, following changes are done
in addition to fixing the CGCONTRDP parsing
issue.
- As the getter functions introduced to get Default
PDN address is not needed, functions are removed.
- Default PDN context information is stored in
the CChannel_Data. With this m_pPDPListData which
holds the default PDN context information is removed.
- New function ExtractLocalAddressAndSubnetMask is
added to extract local_addr and subnet_mask.
handle the <local_addr and subnet_mask> conversion.

Change-Id: I6b49307b81f21ce76b607c7f8a828751df3b2dcb
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/112960
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>