rapid_ril: Introduce RIL_REQUEST_SET_BAND_MODE in 7160
BZ: 114151

The +XBANDSEL and +XUBANDSEL commands are going to be
deprecated in 7160 modems, and so there is a need to
separate the handling of RIL_REQUEST_SET_BAND_MODE from
te_xmm6260 to te_xmm7160.

This patch separates the handling of RIL_REQUEST_SET_BAND_MODE
to te_xmm7160, but uses the 6260 function till the modem fw is
ready.

Change-Id: I0d160ce9bdb12c470e83e222cb42bb4dac550afc
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/113294
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>