rril: Fix for issue where modem initialization takes long time
BZ: 118954

With the new architecture, init commands are handled
in the silo's. Since the silo's are added in multiple
channels, init commands are sent in all the channels
even if unrelated. As a result of this, device takes
long time to camp.

Following changes are done to fix the issue:
- Update m_szChannelBasicInitCmd with correct basic
initialization commands.
- Concatenate "|" before concatenating the TE specific
init commands.
- Remove adding of silos to multiple channels. Since some
URCs are received in all channels(eg: RING, PBREADY), parsing
functions for those URCs are added in all silo's to ignore
them.
- Move +XCONFIG=3,0 to URC initialization string list.
- If there is no command to send as part of modem
initialization, then call TriggerInitStringCompleteEvent
to inform of init string completion for the given
init index and channel.

Change-Id: I62e65aa8163597249e9e51d0958c163d46d940b2
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/115428
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>