rril: Fix usage of pointers to out of scope stack variables
BZ: 119670

Function 'GetActiveDataCallInfoList' uses local variables to
store the strings in 'RIL' format for reporting to AOSP RIL
in structure RIL_Data_Call_Response_v6.

Those variables are however out of scope when the reporting
is done leading to possible memory corruption.

Moreover, in case multiple network interfaces are active, only
the information for the last one is reported (but multiple
times).

Fix is to use the memory allocated in S_ND_PDP_CONTEXT_DATA
structure for this purpose.

Change-Id: I33973d4db8fb4c1b168ed3edbd5724ff7c9ee286
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Reviewed-on: http://android.intel.com:8080/115962
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>