rril: Fix issue in establishing data call with 6360 modem
BZ: 120883

From 6360, data channel should be decided based on
data profile(default, MMS, SUPL etc). This profile
is used to get the network interface ID. So, 6360
modem should override CoreSetupDataCall and call
GetFreeHsiChanel which is allocating channel based
on profile. Since this is not done, default and MMS
connection end up with same network interface id
and also data is over MUX. Due to this sending and
receiving of MMS doesn't work with 6360 modem.

Override the CoreSetupDataCall function in CTE_XMM6360
modem. Since the CoreSetupDataCall for CTE_XMM6360 and
CTE_XMM7160 are same, remove the overriding done in
CTE_XMM7160 modem.

Change-Id: I5283e5e7ba721153f2277f861ab1fe58e0948154
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/116843
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>