rril: Fix no camp issue seen on flight mode act/deact
BZ: 121407

Currently, acquire modem resource request is sent even
when the resource is already acquired. So, when the modem
resource acquire fails(even when it is already acquired)
RADIO_POWER request is completed with error but no radio
state changed event is sent resulting in device stuck in
no camp state.

Changes done to address this issue:
- Acquire the modem resource only if it is not already
acquired.
- If the resource acquire fails, fake the radio states to
force the framework to trigger the RADIO_POWER request again.
- Complete the RADIO_POWER request only after the radio state
is set and also notified to framework.

Change-Id: Idc3ef3092d4a7a50e8f252dabcfab69b4d4d110d
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/117182
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>