rril: Fix operator selection sent before channel init
BZ: 123273

After modem self-reset, operator selection and
registration command is sent even before the basic
initialisation of all the channels are complete.
Even though this doesn't result in major issue
it is better to address this.

Following changes are done to address this issue
- After cold reset/core dump/warm reset, notify the
radio state as off only after the basic initilisation
of all channels are completed. This will force the
framework to send the RADIO_POWER request based
on the current radio state settings.
- Remove the XSIMSTATE? command sent as part of the
basic channel initialisation list.
- Issue XSIMSTATE? only after the basic initialisation
completion of all the channels.

Change-Id: I1153149bc93b3eb7c7af1f9460cff9836e694bad
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/118493
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>