rril: Fix issue in showing data icon when LTE registered
BZ: 122289

When the device registers to LTE network in screen off
state, all are working fine but the data icon is not
shown in the UI.

When the device is registered to LTE network, registration
information is cached in the m_sEPSStatus. When telephony
framework queries the data registration state, copy the
registration information cached from m_sEPSStatus.

Change-Id: I66ff960a09ec20c1c786ebe04834b0801062459d
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/118197
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Baglin, Matthieu <matthieu.baglin@intel.com>
Tested-by: Baglin, Matthieu <matthieu.baglin@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>