rril: Fix issue in getting sim sms store status
BZ: 122971

SIM SMS store status is read by issuing CPMS?. Current
response timeout value for CPMS? is set to 10seconds.
Since the CPMS? can take more than 10seconds, rapid ril
considers this as no response from modem and starts
retrying the request. Also, CPMS? query is not sent to
modem even when the SIM SMS cache is ready on modem side.
This is because the request is blocked in the rapid ril
itself due to request not allowed in the radio off state
and also +XSIM: 12 won't be received anymore.

Following changes are done to address this issue
- Send CPMS? request in the sim ready state.

Change-Id: Id9e6ec95466de92255f62473469dd84a043c67be
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/118110
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>