rril: Update sim app_type as unknown when no modem
BZ: 123623

Default APN information is not sent because as
per the framework it is already sent. When the
modem is powered off, rapid ril doesn't update
the right app type. So, as per framework there
is no change in the application resulting in
IMSI_READY not sent when the modem is powered
on and the SIM is ready.

In order to force the framework to clean up the
UICC application and SIM records on on airplane
mode activation(FMMO), complete the GET_SIM_STATUS
with APPTYPE set to UNKNOWN. This will force the
framework to cleanup the application and also the
SIM records. So, when the sim is ready after airplane
mode deactivation, IMSI_READY will be notified to
all the clients thus resulting in default APN
information sent to modem correctly.

Change-Id: I513e11f08d51f32df316566fa381e399e837dc00
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/118839
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>