rril: report camped access technology while in call
BZ: 125387

In MO/MT CSFB voice call, wrong access technology is
shown instead of the camped access technology.

Only camped access technology is known during call
and it is available as part of XREG URC. Registered
network's access technology which is reported as part
of CREG, CGREG and CEREG is sent only after the end of
voice call.

In CSFB during voice call, report the camped access
technology to the framework.

Change-Id: I310e7c73ebd55470e1f8f0aac6e99d5505e04f8a
Signed-off-by: Pierre-Yves Struillou <Pierre-yvesX.Struillou@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/121851
Reviewed-by: Struillou, Pierre-yvesX <pierre-yvesx.struillou@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>