rapid_ril: PIN code is requested after exiting to airplane mode
BZ: 129208

After a modem cold reset, a flag is set to mark that the current PIN
is saved and can be reused.
When the phone starts up for the first time, following a cold reset,
the flag is also set but there is no PIN saved to reuse.

As a solution, the flag that marks the presence of a saved PIN is reset
when Pcache_Get_PIN() cannot find a record so the phone will save the
new PIN for reuse.

Change-Id: I030defd6db5790da78a05d416ebc4d2201386446
Signed-off-by: SergiuX Bratu <sergiux.bratu@intel.com>
Reviewed-on: http://android.intel.com:8080/124974
Reviewed-by: Muraru, PetruX <petrux.muraru@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Sarghe, AngelaX <angelax.sarghe@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>