rapid_ril: Cannot activate PDP on user build whereas working on user-debug build
BZ: 132365

FreeATCmd1 and FreeATCmd2 was not done under FullLogBuild
(eng or user-debug) but it was done for user builds. Actually
freeing AT cmd1 was done a bit too early, as it may be used
later in the SendCommand code (inside GetResponse and in case of
error).
This patch removes the check of build type to free the AT
command, as the AT command should be freed in all builds.
The freeing is moved to a safer place, to avoid using the freed
memory.

Change-Id: I2ff59e576eb863e0037aaaa3030008b97e83ffd6
Signed-off-by: William Begoc <williamx.begoc@intel.com>
Reviewed-on: http://android.intel.com:8080/128267
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>