rril: handle well +XSIM: 6 (SIM error)
BZ: 133211

In the current code, the +XSIM: 8 is badly handled. This URC
reports a request error. So it concerns one SIM request. It
is not a SIM card error status. It's the +XSIM: 6 URC that we
must used to report SIM card error status.

The +XSIM: 8 URC must be ignored as this one is not supported
by Android FW.

This patch updates the code to correclty manage the +XSIM: 6
case and remove what was done for +XSIM: 8.

Change-Id: I5683301b88773177630a23c8e6ec71670a6cf3bf
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/128271
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>