rapid_ril: Periodic cell information reporting enhancement
BZ: 133570

A cell info cache is created to store the received cell info
in case of unsolicited cell info reporting. The information
is reported to framework only if there is a change in cell info.

For rate setting 0: Cell information is read whenever there is
change in registration and the read cell info values are different
from what is stored in cache.

For rate setting INT_MAX: No periodic reporting of cell information

For all other rate setting: The minimum rate is 60 seconds in rril.

Change-Id: I2892a69c8c261fd6fe61f679a3178556bbfb7fca
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/129551
Reviewed-by: Bourdos, Sylvain <sylvain.bourdos@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>