rril: disable the PIN caching if the SIM is not ready
BZ: 134477

If we remove the SIM card before to go in air plane mode and
reinsert it during the air plane mode, the PIN code won't be
requested after the air plane mode off.

This issue is due to the fact that we enable the PIN code
caching whatever the SIM card status. This issue could be
also seen if we have an APIMR and if we reinsert the SIM
card during the reboot phase of the modem.

This patch fixes the issue by forcing the disabling of the
PIN code caching if the SIM card is not ready.

Change-Id: Iea642e49da94233f020f99e8fea6d5a325760562
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/130957
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Roulliere, Pierre <pierre.roulliere@intel.com>
Reviewed-by: Goea, Andrei DanielX <andrei.danielx.goea@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>