[REVERT ME] rril: update +XCELLINFO channel
BZ: 139146

The +XCELLINFO? request can blocked the DLC2 channel during
some seconds due to the needed time to have the answer from
modem side. This blocking can have side effect on UI side
due to SET/GET AT commands not executed immediately.

This patch updates the RIL code to use the DLC9 which is
less used than the DLC2 to limit the side effect on the UI.

This patch is a workaround waiting an improvement on modem
side to have the XCELLINFO data coming via an URC.

Change-Id: I9c41ac32275ef95d019b396b81fe8f8e7b159907
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/133301
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Riviere, NathanX <nathanx.riviere@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>