rapid_ril: Fix for wrong icon shown when MT CSFB call
BZ: 140778

The root cause of this issue is that while the device is
in 4G and screen is off, the registration notifications are
disabled. When an incoming call is received, though EPS is
not registered, the data registration shows LTE, because the
access technology change notifications are disabled in screen
off state

This patch issues +CEREG? and +XREG? while registered to EPS
when data registration state is requested by framework.

Change-Id: Id70d99fb6551be156cdc7327dfd8d08994bd94de
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/134126
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>