rril: Dont invoke parsing functions when no command sent
BZ: 140928

For ril requests RIL_REQUEST_GSM_SET_BROADCAST_SMS_CONFIG
and RIL_REQUEST_VOICE_RADIO_TECH, parsing functions are
invoked even when no command is sent to modem.

If there is no command sent for a ril request, then the
parsing functions should not be called.

If there is no command to be sent to modem for processing
a ril request, then complete the request immediately.

Note: Post command handlers will be invoked even when
no command is sent to modem. This is to make sure that
any post processing needed for a ril request is done.

Change-Id: I32d424b1c82378235d2db3e519a5df0bf798c44c
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/134143
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Riviere, NathanX <nathanx.riviere@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>