rril: Pin caching in nvram or file system
BZ: 99430

With this patch, pin caching is now done in nvram or
fs based on the mode set in repository.txt.

If the PinCacheMode is set to nvram, then following
files are used to cache the encrypted pin
- /sys/kernel/modem_nvmram/size
Contains the size of the buffer in bytes

- /sys/kernel/modem_nvmram/dump
Cached UICC ID and encrypted pin is stored in this
zone. Size is limited to buffer size contained
in the /sys/kernel/modem_nvram/size

- /sys/kernel/modem_nvram/clear
Any writing operation will clear the zone.

If the PinCacheMode is set to file system,
then the encrypted pin is cached in /config/telephony/dump.

Caches are cleared on sim removal or if the wakeup
is not due to watchdog rest.

Change-Id: I91d3fe58b4aa49b0b0a99855bff66211e99dc0df
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>