rril: Fix device not registering when conformance property set
BZ: 141820

When the conformance property is set, device is not
registering to network due to sim status change not
notified to framework.

When response for the last basic initialization command
has been received from modem, channel's basic initialization
completion flag will be set. Once all the channel's basic
initialization is complete, basic initialization completion
event will be triggered. Upon basic initialization completion
event, sim status will be queried from the modem and sim
status will be updated to the framework.

When the conformance property is set, an extra | gets
appended to the ATCMD channel's basic initialization
string. Since there is no valid command after |, no
command is sent to the modem and ATCMD channel's basic
initialization completion flag is not set. Even when the
basic initialization for all the channels are done, basic
initialization event is not triggered. This results in sim
status not getting queried from modem and not updated to
the framework. Since the sim is not ready, automatic network
selection command is not sent to the modem resulting in
device not registering to the network.

Concatenate | to the init string only when there is a
command to be concatenated after |.

Change-Id: I65817a350f94845511d66553d521df82f6884ab0
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/135297
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Riviere, NathanX <nathanx.riviere@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>